Cleanup: move common functions from grpc_backend.rs and thrift_backend.rs to util.rs
